+++ android/build.gradle
@@ -330,7 +330,6 @@ task prepareHermes() {
     }
 }
 
-// Create Android.mk library module based on jsc from npm
 task prepareJSC {
     doLast {
         def jscPackagePath = findNodeModulePath(projectDir, "jsc-android")
@@ -460,7 +459,6 @@ repositories {
 
 dependencies {
     //noinspection GradleDynamicVersion
-    // implementation "org.webkit:android-jsc:+"
     implementation "com.facebook.react:react-native:+"  // From node_modules
     implementation "androidx.transition:transition:1.1.0"
     api("com.facebook.fbjni:fbjni-java-only:0.0.3")
