+++ android/src/main/cpp/NativeProxy.cpp
@@ -1,21 +1,12 @@
 #include <memory>
 #include <string>
-#include <fbjni/fbjni.h>
-#include <jsi/jsi.h>
 #include <react/jni/ReadableNativeArray.h>
 #include <react/jni/ReadableNativeMap.h>
 #include <jsi/JSIDynamic.h>
 #include "NativeProxy.h"
 #include "AndroidErrorHandler.h"
 #include "AndroidScheduler.h"
-#include <android/log.h>
 #include "PlatformDepMethodsHolder.h"
-#include <cxxreact/NativeToJsBridge.h>
-#include "Logger.h"
-#include <cxxreact/MessageQueueThread.h>
-#include <cxxreact/SystraceSection.h>
-#include <hermes/hermes.h>
-#include <jsi/decorator.h>
 #include <hermes/inspector/RuntimeAdapter.h>
 #include <hermes/inspector/chrome/Registration.h>
