+++ android/src/main/cpp/headers/NativeProxy.h
@@ -14,54 +14,16 @@
 #include "AndroidScheduler.h"
 #include <cxxreact/MessageQueueThread.h>
 #include <cxxreact/SystraceSection.h>
-#include <hermes/hermes.h>
-#include <jsi/decorator.h>
 #include <hermes/inspector/RuntimeAdapter.h>
 #include <hermes/inspector/chrome/Registration.h>
 #include "RuntimeProvider.h"
+#include "HermesExecutorRuntimeAdapter.h"
 
 namespace reanimated {
 
 class RuntimeProvider;
 
 using namespace facebook;
-
-class HermesExecutorRuntimeAdapter : public facebook::hermes::inspector::RuntimeAdapter {
- public:
-  HermesExecutorRuntimeAdapter(
-      std::shared_ptr<facebook::jsi::Runtime> runtime,
-      facebook::hermes::HermesRuntime &hermesRuntime,
-      std::shared_ptr<MessageQueueThread> thread)
-      : runtime_(runtime),
-        hermesRuntime_(hermesRuntime),
-        thread_(std::move(thread)) {}
-
-  virtual ~HermesExecutorRuntimeAdapter() = default;
-
-  facebook::jsi::Runtime &getRuntime() override {
-    return *runtime_;
-  }
-
-  facebook::hermes::debugger::Debugger &getDebugger() override {
-    return hermesRuntime_.getDebugger();
-  }
-
-  void tickleJs() override {
-    // The queue will ensure that runtime_ is still valid when this
-    // gets invoked.
-    thread_->runOnQueue([&runtime = runtime_]() {
-      auto func =
-          runtime->global().getPropertyAsFunction(*runtime, "__tickleJs");
-      func.call(*runtime);
-    });
-  }
-
- public:
-  std::shared_ptr<facebook::jsi::Runtime> runtime_;
-  facebook::hermes::HermesRuntime &hermesRuntime_;
-
-  std::shared_ptr<MessageQueueThread> thread_;
-};
 
 class AnimationFrameCallback : public HybridClass<AnimationFrameCallback> {
  public:
@@ -160,8 +122,5 @@ class NativeProxy : public jni::HybridClass<NativeProxy> {
       std::shared_ptr<facebook::react::CallInvoker> jsCallInvoker,
       std::shared_ptr<Scheduler> scheduler);
 };
-
-
-
 
 }
