+++ android/src/main/cpp/headers/RuntimeProvider.h
@@ -3,12 +3,10 @@
 #include <memory>
 #include <fbjni/fbjni.h>
 #include <jsi/jsi.h>
-#include <jsi/JSIDynamic.h>
 #include <cxxreact/NativeToJsBridge.h>
 #include <cxxreact/MessageQueueThread.h>
 #include <cxxreact/SystraceSection.h>
 #include "NativeProxy.h"
-#include <jsi/decorator.h>
 #include <hermes/inspector/RuntimeAdapter.h>
 #include <hermes/inspector/chrome/Registration.h>
