+++ package.json
@@ -57,7 +57,6 @@
   "dependencies": {
     "@babel/plugin-transform-object-assign": "^7.10.4",
     "fbjs": "^3.0.0",
-    "jsc-android": "^250230.2.1",
     "mockdate": "^3.0.2",
     "string-hash-64": "^1.0.3"
   },
